requirements.txt
python-dotenv==1.0.1
pandas==2.2.3
langchain-core==0.3.29
langchain-openai==0.2.14
langchain-community==0.3.14
langchain-chroma==0.2.0
langchain-ollama==0.2.2
arxiv==2.1.3
pypdf==5.1.0
scikit-learn==1.6.1
seaborn==0.13.2
chromadb==0.5.23
pydantic==2.10.5
openai==1.59.6
tiktoken==0.8.0
numpy==1.26.2
tqdm==4.67.1